+++ src/components/Company.js
